Refactor NormNet and related classes to use long data type for values and weights

+++ src/main/java/de/schmiereck/prednet/service/normNet/NormSynapse.java
@@ -2,10 +2,10 @@ package de.schmiereck.prednet.service.normNet;
 
 public class NormSynapse {
     public final NormNeuron parentNeuron;
-    public int weight;
+    public long weight;
     public final NormNeuron childNeuron;
 
-    public NormSynapse(final NormNeuron parentNeuron, final int weight, final NormNeuron childNeuron) {
+    public NormSynapse(final NormNeuron parentNeuron, final long weight, final NormNeuron childNeuron) {
         this.parentNeuron = parentNeuron;
         this.weight = weight;
         this.childNeuron = childNeuron;
